+++ app/graphql/context.js
@@ -1,4 +1,3 @@
-import jwt from 'jsonwebtoken'
 import { getAuth, getRequestingGroup, getRequestingService } from '../auth/authenticate.js'
 import { config } from '../config.js'
 import { HitachiPayments } from '../data-sources/hitachi/HitachiPayments.js'
@@ -9,25 +8,8 @@ import { RuralPaymentsBusiness } from '../data-sources/rural-payments/RuralPayme
 import { RuralPaymentsCustomer } from '../data-sources/rural-payments/RuralPaymentsCustomer.js'
 import { RuralPaymentsReferenceData } from '../data-sources/rural-payments/RuralPaymentsReferenceData.js'
 import { Permissions } from '../data-sources/static/permissions.js'
-import { BadRequest } from '../errors/graphql.js'
 import { logger } from '../logger/logger.js'
 import { db } from '../mongo.js'
-
-export const extractOrgIdFromDefraIdToken = (sbi, token) => {
-  const { payload } = jwt.decode(token, { complete: true })
-  if (payload?.relationships && Array.isArray(payload.relationships)) {
-    // Find relationship string that matches the given SBI
-    const relationship = payload.relationships.find((rel) => {
-      const [, tokenSBI] = rel.split(':')
-      return sbi === tokenSBI
-    })
-    if (relationship) {
-      const [orgId] = relationship.split(':')
-      return orgId
-    }
-  }
-  throw new BadRequest('Defra ID token is not valid for the provided SBI')
-}
 
 function stripClientSuppliedServiceAccountHeader(request) {
   // Service account foundations have been added, to support the DAL internal service account, but this is not
